+++ src/Pim/Bundle/CatalogBundle/Entity/Channel.php
@@ -17,17 +17,13 @@ use Pim\Bundle\CatalogBundle\Model\ReferableInterface;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @Config(
- *  defaultValues={
- *      "entity"={"label"="Channel", "plural_label"="Channels"},
- *      "security"={
- *          "type"="ACL",
- *          "group_name"=""
- *      }
- *  }
- * )
- *
  * @ExclusionPolicy("all")
+ *
+ * @Config(
+ *     defaultValues={
+ *         "entity"={"label"="Channel", "plural_label"="Channels"}
+ *     }
+ * )
  */
 class Channel implements ReferableInterface
 {