+++ src/Pim/Bundle/CatalogBundle/Form/Handler/ProductAttributeHandler.php
@@ -8,7 +8,7 @@ use Doctrine\Common\Persistence\ObjectManager;
 use Pim\Bundle\CatalogBundle\Model\ProductAttributeInterface;
 use Pim\Bundle\CatalogBundle\Entity\AttributeOption;
 use Pim\Bundle\CatalogBundle\Entity\AttributeOptionValue;
-use Pim\Bundle\CatalogBundle\Manager\AttributeTypeManager;
+use Pim\Bundle\CatalogBundle\Manager\ProductAttributeManagerInterface;
 
 /**
  * Form handler for Product attribute
@@ -35,27 +35,27 @@ class ProductAttributeHandler
     protected $manager;
 
     /**
-     * @var AttributeTypeManager
+     * @var ProductAttributeManagerInterface
      */
-    protected $attTypeManager;
+    protected $attributeManager;
 
     /**
      * Constructor for handler
-     * @param FormInterface        $form           Form called
-     * @param Request              $request        Web request
-     * @param ObjectManager        $manager        Storage manager
-     * @param AttributeTypeManager $attTypeManager Attribute type manager
+     * @param FormInterface                    $form             Form called
+     * @param Request                          $request          Web request
+     * @param ObjectManager                    $manager          Storage manager
+     * @param ProductAttributeManagerInterface $attributeManager Attribute type manager
      */
     public function __construct(
         FormInterface $form,
         Request $request,
         ObjectManager $manager,
-        AttributeTypeManager $attTypeManager
+        ProductAttributeManagerInterface $attributeManager
     ) {
         $this->form    = $form;
         $this->request = $request;
         $this->manager = $manager;
-        $this->attTypeManager = $attTypeManager;
+        $this->attributeManager = $attributeManager;
     }
 
     /**
@@ -64,11 +64,11 @@ class ProductAttributeHandler
      */
     public function preProcess($data)
     {
-        $attribute = $this->attTypeManager->createAttributeFromFormData($data);
+        $attribute = $this->attributeManager->createAttributeFromFormData($data);
 
         $this->form->setData($attribute);
 
-        $data = $this->attTypeManager->prepareFormData($data);
+        $data = $this->attributeManager->prepareFormData($data);
 
         $this->form->bind($data);
     }
@@ -177,7 +177,7 @@ class ProductAttributeHandler
             }
         }
 
-        $this->attTypeManager->prepareBackendProperties($entity);
+        $this->attributeManager->prepareBackendProperties($entity);
 
         $this->manager->persist($entity);
         $this->manager->flush();