+++ src/Pim/Bundle/ImportExportBundle/Transformer/ORMAttributeTransformer.php
@@ -4,7 +4,7 @@ namespace Pim\Bundle\ImportExportBundle\Transformer;
 
 use Symfony\Bridge\Doctrine\RegistryInterface;
 use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
-use Pim\Bundle\CatalogBundle\Manager\ProductManager;
+use Pim\Bundle\CatalogBundle\Manager\ProductAttributeManagerInterface;
 use Pim\Bundle\ImportExportBundle\Transformer\ColumnInfo\ColumnInfoTransformerInterface;
 use Pim\Bundle\ImportExportBundle\Transformer\Guesser\GuesserInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductAttributeInterface;
@@ -20,9 +20,9 @@ use Pim\Bundle\ImportExportBundle\Cache\EntityCache;
 class ORMAttributeTransformer extends ORMTransformer
 {
     /**
-     * @var ProductManager
+     * @var ProductAttributeManagerInterface
      */
-    protected $productManager;
+    protected $attributeManager;
 
     /**
      * @var EntityCache
@@ -32,22 +32,23 @@ class ORMAttributeTransformer extends ORMTransformer
     /**
      * Constructor
      *
-     * @param RegistryInterface              $doctrine
-     * @param PropertyAccessorInterface      $propertyAccessor
-     * @param GuesserInterface               $guesser
-     * @param ColumnInfoTransformerInterface $columnInfoTransformer
-     * @param ProductManager                 $productManager
+     * @param RegistryInterface                $doctrine
+     * @param PropertyAccessorInterface        $propertyAccessor
+     * @param GuesserInterface                 $guesser
+     * @param ColumnInfoTransformerInterface   $columnInfoTransformer
+     * @param ProductAttributeManagerInterface $attributeManager
+     * @param EntityCache                      $entityCache
      */
     public function __construct(
         RegistryInterface $doctrine,
         PropertyAccessorInterface $propertyAccessor,
         GuesserInterface $guesser,
         ColumnInfoTransformerInterface $columnInfoTransformer,
-        ProductManager $productManager,
+        ProductAttributeManagerInterface $attributeManager,
         EntityCache $entityCache
     ) {
         parent::__construct($doctrine, $propertyAccessor, $guesser, $columnInfoTransformer);
-        $this->productManager = $productManager;
+        $this->attributeManager = $attributeManager;
         $this->entityCache = $entityCache;
     }
 
@@ -56,7 +57,7 @@ class ORMAttributeTransformer extends ORMTransformer
      */
     protected function setProperties($class, $entity, array $data)
     {
-        $attributeClass = $this->productManager->getAttributeName();
+        $attributeClass = $this->attributeManager->getAttributeClass();
         if ($attributeClass === $class && isset($data['options'])) {
             $optionsData = $data['options'];
             unset($data['options']);
@@ -81,7 +82,7 @@ class ORMAttributeTransformer extends ORMTransformer
             if (!isset($optionData['code'])) {
                 $optionData['code'] = $code;
             }
-            $attributeClass = $this->productManager->getAttributeOptionName();
+            $attributeClass = $this->attributeManager->getAttributeOptionClass();
             $option = $this->getEntity($attributeClass, $optionData);
             $this->setProperties($attributeClass, $option, $optionData);
             if (count($this->errors)) {
@@ -97,8 +98,8 @@ class ORMAttributeTransformer extends ORMTransformer
      */
     protected function createEntity($class, array $data)
     {
-        return ($this->productManager->getAttributeName() === $class)
-            ? $this->productManager->createAttribute($data['type'])
+        return ($this->attributeManager->getAttributeClass() === $class)
+            ? $this->attributeManager->createAttribute($data['type'])
             : parent::createEntity($class, $data);
     }
 }