+++ src/Pim/Bundle/InstallerBundle/FixtureLoader/ConfigurationRegistry.php
@@ -110,6 +110,7 @@ class ConfigurationRegistry implements ConfigurationRegistryInterface
      * Return a fixture service
      *
      * @param string $service
+     * @param string $name
      * @param string $extension
      *
      * @return object